refactor: models에 queries, service로 쿼리옵션용 폴더 추가
