+++ README.md
@@ -2,7 +2,7 @@
 
 - assets : 특정 컴포넌트에 종속되 있는 모듈단위의 파일들 (주로 js/ts, css, 이미지모듈등) - 근데 미사용으로 보임
 - components
-  - feature : 비즈니스 도메인과 관련된 컴포넌트들
+  - feature : 비즈니스 도메인과 관련된 컴포넌트들 (관심사별로 구분)
   - layout : 도메인과 관련없는 레이아웃? (아직 미정)
   - shared : 공용 컴포넌트들
 - constants : 상수
