+++ src/components/shared/Floating/Floating.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import { cn } from '@/lib/cn';
 import useModal from '@/store/use-modal-store';
-import useGetMe from '@/models/auth/useGetMe';
+import useGetMe from '@/models/queries/auth/useGetMe';
 
 const Floating = () => {
   const { onOpen } = useModal();
