+++ src/components/shared/modals/item-edit-modal.tsx
@@ -12,7 +12,7 @@ import axios from 'axios';
 import apiInstance from '@/lib/axios';
 import useModal from '@/store/use-modal-store';
 
-import useGetProductDetail from '@/models/product/useGetProductDetail';
+import useGetProductDetail from '@/models/queries/product/useGetProductDetail';
 import DropDown from '../DropDown/DropDown';
 import TextFieldInput from '../Input/TextFieldInput';
 import ImageInput from '../Input/ImageInput';
