+++ src/components/shared/modals/review-edit-modal.tsx
@@ -11,7 +11,7 @@ import Image from 'next/image';
 import { Rating } from 'react-simple-star-rating';
 import useModal from '@/store/use-modal-store';
 
-import useEditReview from '@/models/reviews/useEditReview';
+import useEditReview from '@/models/queries/reviews/useEditReview';
 import ImageInput from '../Input/ImageInput';
 import Button from '../Button/Button';
 import TextAreaInput from '../Input/TextAreaInput';
