+++ src/models/queries/product/useFavoriteProduct.ts
