+++ src/models/queries/product/useGetSortedProducts.ts
@@ -1,27 +1,7 @@
-import { isServer, queryOptions, useQueries } from '@tanstack/react-query';
-import axios from '@/lib/axios';
+import { useQueries } from '@tanstack/react-query';
 import sortConverter from '@/utils/sortConverter';
 import { ORDER_VARIANTS } from '@/constants/products';
-
-export const sortedProductsService = {
-  queryKey: ['sortedProducts'],
-  queryOptions: (sortOrder: string) =>
-    queryOptions({
-      queryKey: ['sortedProduct', sortOrder],
-      queryFn: async () => {
-        const requestUri = isServer
-          ? 'https://mogazoa-api.vercel.app/5-5/products?'
-          : 'products?';
-        const { data } = await axios.get(`${requestUri}order=${sortOrder}`);
-        return {
-          list: data.list,
-          nextCursor: data.nextCursor,
-          sortBy: sortOrder,
-        };
-      },
-      staleTime: 60 * 1000 * 10,
-    }),
-};
+import sortedProductsService from '@/models/services/product/sortedProductsService';
 
 const useGetSortedProducts = () => {
   const orderVariants = ORDER_VARIANTS.map((item) => sortConverter(item));
