+++ src/utils/isServer.ts
@@ -1,5 +1,3 @@
-const isServer = () => {
-  return typeof window !== 'undefined';
-};
+const isServer = typeof window === 'undefined';
 
 export default isServer;
